Configure User Admin and Sign up form - Merged User and UserProfile in admin area - Moved login view to account app -- Changed urls to reflect above
