+++ bapccanada/account/admin.py
@@ -1,6 +1,57 @@
 from django.contrib import admin
 
 # Register your models here.
-from account.models import UserProfile
+from django.contrib.auth.admin import UserAdmin
+from django.contrib.auth.models import User
 
-admin.site.register(UserProfile)
+from account.forms import SignUpForm
+
+# class UserProfileAdmin(admin.ModelAdmin):
+#     list_display = ('user', 'birth_date')
+#     readonly_fields = ('birth_date',)
+#
+#
+# admin.site.register(UserProfile, UserProfileAdmin)
+
+
+# class ProfileInline(admin.StackedInline):
+#     model = UserProfile
+#     can_delete = False
+#     verbose_name_plural = 'Profile'
+#     fk_name = 'user'
+#
+#
+# class CustomUserAdmin(UserAdmin):
+#     inlines = (ProfileInline,)
+#
+#     def get_inline_instances(self, request, obj=None):
+#         if not obj:
+#             return list()
+#         return super(CustomUserAdmin, self).get_inline_instances(request, obj)
+#
+#
+# admin.site.unregister(User)
+# admin.site.register(User, CustomUserAdmin)
+
+
+from .models import UserProfile
+
+
+class ProfileInline(admin.StackedInline):
+    model = UserProfile
+    verbose_name_plural = 'Profile'
+    fk_name = 'user'
+
+
+class CustomUserAdmin(UserAdmin):
+    # list_display = ('user', 'birth_date')
+    inlines = (ProfileInline,)
+
+    def get_inline_instances(self, request, obj=None):
+        if not obj:
+            return list()
+        return super(CustomUserAdmin, self).get_inline_instances(request, obj)
+
+
+admin.site.unregister(User)
+admin.site.register(User, CustomUserAdmin)
