+++ bapccanada/account/urls.py
@@ -15,7 +15,6 @@ urlpatterns = [
     path('user/builds/', views.view_builds, name="builds"),
     path('user/notifications/', views.view_notifications, name="notifications"),
     path('user/security/', views.view_security, name="security"),
-
-
-    path('signup/', UserCreate.as_view(), name='sign_up')
+    path('signup/', UserCreate.as_view(), name='sign_up'),
+    re_path(r'^login/$', views.login_view, name="sign_in"),
 ]
