+++ bapccanada/home/templates/login.html
@@ -1,7 +1,7 @@
 {% extends 'base.html' %}
 {% block bapc-header-title %}{{ title }}{% endblock %}
 {% block bapc-body-content %}
-    <form class="bapc-form-signin" action="{% url 'home:login' %}" method="post">
+    <form class="bapc-form-signin" action="{% url 'account:sign_in' %}" method="post">
         <h1 class="h3 mb-3 font-weight-normal">Please sign in</h1>
 
         {% csrf_token %}
