+++ bapccanada/home/urls.py
@@ -7,5 +7,4 @@ from . import views
 app_name = 'home'
 urlpatterns = [
     re_path(r'^$', HomeView.as_view(), name='home'),
-    re_path(r'^login/$', views.login_view, name="login"),
 ]
