+++ bapccanada/templates/navbar.html
@@ -1,7 +1,4 @@
 {% load static %}
-{% url 'home:login' as login_url %}
-{% url 'account:profile' as profile_url %}
-{% url 'build:create' as build_url %}
 <nav class="navbar navbar-expand-lg navbar-light bg-light">
     <div class="container">
         <a class="navbar-brand" href="/">
@@ -22,7 +19,7 @@
 
             {# TODO Login Configuration #}
             <div class="navbar-account ml-auto">
-                <a class="btn btn-primary" href="{% url 'home:login' %}">Sign In</a>
+                <a class="btn btn-primary" href="{% url 'account:sign_in' %}">Sign In</a>
                 <a class="btn btn-danger" href="{% url 'account:sign_up' %}">Sign Up</a>
                 <button class="btn btn-secondary">Sign Out</button>
                 <a href="{% url 'account:profile' %}">@USER Settings</a>
